lint: apply clippy's machine-applicable fixes across the workspace
All of these come from `cargo clippy --fix`, so each is clippy's own rewrite
rather than a hand edit. The bulk is the poisoned-lock idiom
(`unwrap_or_else(|e| e.into_inner())` -> `unwrap_or_else(PoisonError::into_inner)`,
791 sites), plus must_use attributes, raw-string hashes, and redundant closures.

Behavior is unchanged: the only semantic-looking hunks are added parentheses
where clippy flagged non-obvious precedence, and Rust's existing binding was
already what the parentheses now spell out.

3508 lib tests across 33 crates pass, plus hl-daemon's 209.

+++ src/containers/hl-engine/src/ffi/linux/execution/path/filesystem.rs
@@ -41,18 +41,18 @@ impl HostFilesystem {
         let status = unsafe { status.assume_init() };
         let (kind, zero) = Self::kind(guest);
         let restricted = kind != FilesystemKind::Overlay;
-        let filesystem_id = status.f_fsid as u64;
+        let filesystem_id = status.f_fsid;
         FilesystemStats {
             kind,
-            block_size: status.f_bsize as u64,
-            blocks: if zero { 0 } else { status.f_blocks as u64 },
-            blocks_free: if zero { 0 } else { status.f_bfree as u64 },
-            blocks_available: if zero { 0 } else { status.f_bavail as u64 },
-            files: if zero { 0 } else { status.f_files as u64 },
-            files_free: if zero { 0 } else { status.f_ffree as u64 },
+            block_size: status.f_bsize,
+            blocks: if zero { 0 } else { status.f_blocks },
+            blocks_free: if zero { 0 } else { status.f_bfree },
+            blocks_available: if zero { 0 } else { status.f_bavail },
+            files: if zero { 0 } else { status.f_files },
+            files_free: if zero { 0 } else { status.f_ffree },
             filesystem_id: [filesystem_id as u32, (filesystem_id >> 32) as u32],
             name_maximum: 255,
-            fragment_size: status.f_bsize as u64,
+            fragment_size: status.f_bsize,
             read_only: false,
             nosuid: restricted,
             nodev: restricted,
